on-boarding folder update
