+++ lib/onboarding/onboarding.dart
@@ -1,0 +1,3 @@
+export 'view/start.dart';
+export 'view/second.dart';
+export 'view/third.dart';
