+++ lib/onboarding/view/second.dart
@@ -1,58 +1,57 @@
-import 'package:talao/app/pages/on_boarding/second.dart';
 import 'package:talao/app/pages/on_boarding/tos.dart';
+import 'package:talao/app/shared/router.dart';
 import 'package:talao/app/shared/widget/base/button.dart';
 import 'package:talao/app/shared/widget/base/page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:talao/onboarding/onboarding.dart';
 
-class OnBoardingStartPage extends StatefulWidget {
-  static Route route() {
-    return MaterialPageRoute<void>(
-      builder: (context) => OnBoardingStartPage(),
-    );
-  }
+class OnBoardingSecondPage extends StatefulWidget {
+  static Route route() => RightToLeftRoute(
+        builder: (context) => OnBoardingSecondPage(),
+      );
 
   @override
-  State<OnBoardingStartPage> createState() => _OnBoardingStartPageState();
+  State<OnBoardingSecondPage> createState() => _OnBoardingSecondPageState();
 }
 
-class _OnBoardingStartPageState extends State<OnBoardingStartPage> {
+class _OnBoardingSecondPageState extends State<OnBoardingSecondPage> {
   var animate = true;
 
   @override
   Widget build(BuildContext context) {
     final localizations = AppLocalizations.of(context)!;
-    return WillPopScope(
-      onWillPop: () async => false,
-      child: GestureDetector(
-        onTap: () async {
-          await Navigator.of(context).push(OnBoardingSecondPage.route());
-        },
-        onHorizontalDragUpdate: (drag) async {
-          if (animate && drag.delta.dx < -2) {
+    return GestureDetector(
+      onHorizontalDragUpdate: (drag) async {
+        if (animate) {
+          if (drag.delta.dx > 2) {
+            Navigator.of(context).pop();
             disableAnimation();
-            await Navigator.of(context).push(
-              OnBoardingSecondPage.route(),
-            );
           }
-        },
-        child: BasePage(
-          scrollView: true,
+
+          if (drag.delta.dx < -2) {
+            disableAnimation();
+            await Navigator.of(context).push(OnBoardingThirdPage.route());
+          }
+        }
+      },
+      child: BasePage(
           backgroundColor: Theme.of(context).colorScheme.surface,
+          scrollView: true,
           body: Column(
             mainAxisSize: MainAxisSize.min,
             children: <Widget>[
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: Image.asset(
-                  'assets/image/slide_1.png',
+                  'assets/image/slide_2.png',
                   height: 200,
                 ),
               ),
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: Text(
-                  localizations.appPresentation1,
+                  localizations.appPresentation2,
                   style: Theme.of(context).textTheme.bodyText1,
                   textAlign: TextAlign.justify,
                 ),
@@ -79,16 +78,16 @@ class _OnBoardingStartPageState extends State<OnBoardingStartPage> {
                         children: [
                           Icon(
                             Icons.circle,
-                            color: Theme.of(context).colorScheme.secondary,
+                            color: Theme.of(context)
+                                .colorScheme
+                                .secondary
+                                .withOpacity(0.2),
                             size: 15,
                           ),
                           SizedBox(width: 10),
                           Icon(
                             Icons.circle,
-                            color: Theme.of(context)
-                                .colorScheme
-                                .secondary
-                                .withOpacity(0.2),
+                            color: Theme.of(context).colorScheme.secondary,
                             size: 15,
                           ),
                           SizedBox(width: 10),
@@ -117,9 +116,7 @@ class _OnBoardingStartPageState extends State<OnBoardingStartPage> {
                 ),
               ),
             ),
-          ),
-        ),
-      ),
+          )),
     );
   }
